# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/internal/simbalance/update": {
      "put": {
        "tags": [
          "sim-balance-internal-controller"
        ],
        "operationId": "update_5",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSimBalanceDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SimBalanceDto"
                }
              }
            }
          }
        }
      }
    },
    "/internal/sim/update": {
      "put": {
        "tags": [
          "sim-internal-controller"
        ],
        "operationId": "update_6",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSimDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SimDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/simbalance/create": {
      "post": {
        "tags": [
          "sim-balance-internal-controller"
        ],
        "operationId": "create_5",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSimBalanceDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SimBalanceDto"
                }
              }
            }
          }
        }
      }
    },
    "/internal/sim/create": {
      "post": {
        "tags": [
          "sim-internal-controller"
        ],
        "operationId": "create_6",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSimDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SimDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/simbalance/page": {
      "get": {
        "tags": [
          "sim-balance-internal-controller"
        ],
        "operationId": "page_5",
        "parameters": [
          {
            "name": "dto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SearchSimBalanceDto"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDtoSimBalanceDto"
                }
              }
            }
          }
        }
      }
    },
    "/internal/simbalance/list": {
      "get": {
        "tags": [
          "sim-balance-internal-controller"
        ],
        "operationId": "list",
        "parameters": [
          {
            "name": "dto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SearchSimBalanceDto"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Sort"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimBalanceDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/internal/sim/page": {
      "get": {
        "tags": [
          "sim-internal-controller"
        ],
        "operationId": "page_6",
        "parameters": [
          {
            "name": "dto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SearchSimDto"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDtoSimDto"
                }
              }
            }
          }
        }
      }
    },
    "/internal/sim/list": {
      "get": {
        "tags": [
          "sim-internal-controller"
        ],
        "operationId": "list_1",
        "parameters": [
          {
            "name": "dto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SearchSimDto"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Sort"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/simbalance/{id}/delete": {
      "delete": {
        "tags": [
          "sim-balance-internal-controller"
        ],
        "operationId": "delete_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/internal/sim/{id}/delete": {
      "delete": {
        "tags": [
          "sim-internal-controller"
        ],
        "operationId": "delete_6",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Operator": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "DELETED"
            ]
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateSimBalanceDto": {
        "required": [
          "createDate",
          "id",
          "simId",
          "value"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор",
            "format": "uuid"
          },
          "value": {
            "type": "integer",
            "description": "Баланс",
            "format": "int32"
          },
          "simId": {
            "type": "string",
            "description": "id Sim-карты",
            "format": "uuid"
          },
          "createDate": {
            "type": "string",
            "description": "Дата создания",
            "format": "date-time"
          }
        }
      },
      "SimBalanceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор",
            "format": "uuid"
          },
          "value": {
            "type": "integer",
            "description": "Баланс",
            "format": "int32"
          },
          "sim": {
            "$ref": "#/components/schemas/SimDto"
          },
          "createDate": {
            "type": "string",
            "description": "Дата создания",
            "format": "date-time"
          }
        },
        "description": "DTO баланса Sim-карты"
      },
      "SimDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор",
            "format": "uuid"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Номер телефона"
          },
          "operator": {
            "$ref": "#/components/schemas/Operator"
          }
        },
        "description": "DTO Sim-карты"
      },
      "UpdateSimDto": {
        "required": [
          "id",
          "operatorId",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор",
            "format": "uuid"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Номер телефона"
          },
          "operatorId": {
            "type": "string",
            "description": "id оператора",
            "format": "uuid"
          }
        },
        "description": "DTO обновления Sim-карты"
      },
...
      "CreateSimBalanceDto": {
        "required": [
          "createDate",
          "simId",
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "integer",
            "description": "Баланс",
            "format": "int32"
          },
          "simId": {
            "type": "string",
            "description": "id Sim-карты",
            "format": "uuid"
          },
          "createDate": {
            "type": "string",
            "description": "Дата создания",
            "format": "date-time"
          }
        }
      },
      "CreateSimDto": {
        "required": [
          "operatorId",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Номер телефона"
          },
          "operatorId": {
            "type": "string",
            "description": "id оператора",
            "format": "uuid"
          }
        },
        "description": "DTO создания Sim-карты"
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Sort": {
        "type": "object",
        "properties": {
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "SearchSimBalanceDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "integer",
            "description": "Баланс",
            "format": "int32"
          },
          "simId": {
            "type": "string",
            "description": "id Sim-карты",
            "format": "uuid"
          },
          "createDateFrom": {
            "type": "string",
            "description": "Дата создания начало",
            "format": "date-time"
          },
          "createDateTo": {
            "type": "string",
            "description": "Дата создания конец",
            "format": "date-time"
          }
        }
      },
      "CollectionDtoSimBalanceDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SimBalanceDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "SearchSimDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Номер телефона"
          },
          "operatorId": {
            "type": "string",
            "description": "id оператора",
            "format": "uuid"
          }
        },
        "description": "DTO поиска Sim-карты"
      },
      "CollectionDtoSimDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SimDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}